pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soma"
version = "0.1.0"
description = "Biometric environment controller — WHOOP + smart home automation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ruthvik"}
]
keywords = ["whoop", "smart-home", "biometrics", "automation", "circadian"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]
dependencies = [
    "python-kasa>=0.7.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "pytz>=2024.1",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
soma-scheduler = "soma.scheduler:main"
soma-dashboard = "soma.server:main"
soma-light = "soma.light_control:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
